requirements.txt
requests
python-dotenv
python-binance
tenacity
flask